Converted Deck to Stack
Converted Deck of Cards into a Stack of cards with a Style of Position
absolute with a width of SCREEN_WIDTH.
applied this styles to current card and the list of cards behind it.
*** We Used reverse() function on map such that the stack shows top
element on top.
*** We used zIndex :1 property in Styles object to overcome the Gotcha
in Android Platforms

After this We Added little bit more styling to stack of cards such that
they show up on top like a pack of cards placed one above the other. we
did this with the help of style { top : 10 * (i - this.state.index) } to
the bottom cards such that they align nicely